requirements.txt
streamlit
pandas
plotly
pdfplumber
python-docx
plotly
scikit-learn
nltk